+++ src/syn/sequenceof.rs
@@ -10,8 +10,8 @@ impl<C: Constraint> Default for SequenceOf<C> {
 }
 
 pub trait Constraint {
-    const MIN: Option<usize> = None;
-    const MAX: Option<usize> = None;
+    const MIN: Option<u64> = None;
+    const MAX: Option<u64> = None;
 }
 
 #[derive(Default)]
